requirements.txt
torch
torchvision
numpy
Pillow
datasets
huggingface-hub
tqdm
matplotlib
seaborn
torch-summary